smart-ncert/requirements.txt
Flask
Flask-Session
google-generativeai>=0.5.0
PyPDF2
langdetect
python-dotenv
flask-cors
